+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * (Android) (bug fix) : sometimes could not run long scripts
 * (bug fix) : script types in some errors were not precise enough : sometimes we missed the piece kind
+* (bug fix) : script error "parenthesis without expression" made program crash instead of reporting it
 
 ## 4.0.148